Fix create-route script and move the java sources to a more standard org.grails.plugin package.
