+++ src/java/org/grails/plugin/routing/DefaultGrailsRouteClass.java
@@ -1,0 +1,14 @@
+package org.grails.plugin.routing;
+
+import org.codehaus.groovy.grails.commons.AbstractInjectableGrailsClass;
+
+@SuppressWarnings("rawtypes")
+public class DefaultGrailsRouteClass extends AbstractInjectableGrailsClass implements GrailsRouteClass, GrailsRouteClassProperty {
+    public DefaultGrailsRouteClass(Class clazz) {
+        super(clazz, ROUTE);
+    }
+
+    public DefaultGrailsRouteClass(Class clazz, String trailingName) {
+        super(clazz, trailingName);
+    }
+}
