+++ src/java/org/grails/plugin/routing/GrailsRouteClass.java
@@ -1,0 +1,6 @@
+package org.grails.plugin.routing;
+
+import org.codehaus.groovy.grails.commons.InjectableGrailsClass;
+
+public interface GrailsRouteClass extends InjectableGrailsClass {
+}
