+++ src/java/org/grails/plugin/routing/GrailsRouteClassProperty.java
@@ -1,0 +1,5 @@
+package org.grails.plugin.routing;
+
+public interface GrailsRouteClassProperty {
+    String ROUTE = "Route";
+}
