--- src/java/org/grails/plugins/routing/GrailsRouteClass.java
@@ -1,6 +1,0 @@
-package org.grails.plugins.routing;
-
-import org.codehaus.groovy.grails.commons.InjectableGrailsClass;
-
-public interface GrailsRouteClass extends InjectableGrailsClass {
-}
