--- src/java/org/grails/plugins/routing/GrailsRouteClassProperty.java
@@ -1,5 +1,0 @@
-package org.grails.plugins.routing;
-
-public interface GrailsRouteClassProperty {
-    String ROUTE = "Route";
-}
